feat: small ui fixes in the dashboard
