+++ dashboard/src/components/screencast/ScreencastTile.tsx
@@ -1,6 +1,9 @@
 import { useCallback, useEffect, useRef, useState } from "react";
 import { addTokenToUrl } from "../../services/auth";
 import * as api from "../../services/api";
+import ScreencastStatusBar, {
+  type ScreencastStatus,
+} from "./ScreencastStatusBar";
 
 interface Props {
   instanceId: string;
@@ -12,8 +15,6 @@ interface Props {
   fps?: number;
   showTitle?: boolean;
 }
-
-type Status = "connecting" | "streaming" | "error";
 
 export default function ScreencastTile({
   instanceId,
@@ -27,12 +28,10 @@ export default function ScreencastTile({
 }: Props) {
   const canvasRef = useRef<HTMLCanvasElement>(null);
   const socketRef = useRef<WebSocket | null>(null);
-  const [status, setStatus] = useState<Status>("connecting");
+  const [status, setStatus] = useState<ScreencastStatus>("connecting");
   const [fpsDisplay, setFpsDisplay] = useState("—");
   const [sizeDisplay, setSizeDisplay] = useState("—");
   const [localFps, setLocalFps] = useState(fps);
-  const [isCapturing, setIsCapturing] = useState(false);
-  const [isPdfGenerating, setIsPdfGenerating] = useState(false);
   const [hasFrame, setHasFrame] = useState(false);
   const [fallbackUrl, setFallbackUrl] = useState<string | null>(null);
   const [retryKey, setRetryKey] = useState(0);
@@ -119,48 +118,6 @@ export default function ScreencastTile({
       }
     };
   }, [fallbackUrl]);
-
-  const takeScreenshot = async () => {
-    if (isCapturing) return;
-    setIsCapturing(true);
-
-    try {
-      const blob = await api.fetchTabScreenshot(tabId, "png");
-      const url = URL.createObjectURL(blob);
-      const link = document.createElement("a");
-      link.href = url;
-      link.download = `screenshot-${tabId}-${Date.now()}.png`;
-      document.body.appendChild(link);
-      link.click();
-      document.body.removeChild(link);
-      URL.revokeObjectURL(url);
-    } catch (e) {
-      console.error("Screenshot capture failed", e);
-    } finally {
-      setIsCapturing(false);
-    }
-  };
-
-  const downloadPdf = async () => {
-    if (isPdfGenerating) return;
-    setIsPdfGenerating(true);
-
-    try {
-      const blob = await api.fetchTabPdf(tabId);
-      const url = URL.createObjectURL(blob);
-      const link = document.createElement("a");
-      link.href = url;
-      link.download = `page-${tabId}-${Date.now()}.pdf`;
-      document.body.appendChild(link);
-      link.click();
-      document.body.removeChild(link);
-      URL.revokeObjectURL(url);
-    } catch (e) {
-      console.error("PDF generation failed", e);
-    } finally {
-      setIsPdfGenerating(false);
-    }
-  };
 
   const captureFallback = useCallback(async () => {
     try {
@@ -288,7 +245,7 @@ export default function ScreencastTile({
         ? "bg-warning"
         : "bg-destructive";
   return (
-    <div className="flex h-full flex-col overflow-hidden rounded-lg border border-border-subtle bg-bg-elevated">
+    <div className="flex h-full flex-col overflow-hidden border border-border-subtle bg-bg-elevated">
       {/* Header */}
       {showTitle && (
         <div className="flex shrink-0 items-center justify-between border-b border-border-subtle px-3 py-2">
@@ -350,61 +307,14 @@ export default function ScreencastTile({
           </div>
         )}
       </div>
-      <div className="flex shrink-0 items-center justify-between border-t border-border-subtle px-3 py-1 text-xs text-text-muted">
-        <div className="flex items-center gap-3">
-          <div className="flex items-center overflow-hidden rounded border border-border-subtle bg-black/20">
-            <button
-              onClick={() => setLocalFps((prev) => Math.max(1, prev - 1))}
-              className="flex h-5 w-5 items-center justify-center hover:bg-white/5 active:bg-white/10"
-              title="Decrease FPS"
-            >
-              -
-            </button>
-            <div className="min-w-16 px-1.5 text-center font-mono text-[10px] text-text-secondary">
-              {localFps} FPS ({fpsDisplay})
-            </div>
-            <button
-              onClick={() => setLocalFps((prev) => Math.min(30, prev + 1))}
-              className="flex h-5 w-5 items-center justify-center hover:bg-white/5 active:bg-white/10"
-              title="Increase FPS"
-            >
-              +
-            </button>
-          </div>
-
-          <button
-            onClick={takeScreenshot}
-            disabled={isCapturing || status !== "streaming"}
-            className={`flex h-6 w-6 items-center justify-center rounded-md border border-border-subtle transition-colors hover:bg-white/5 disabled:opacity-50 ${
-              isCapturing ? "bg-primary/20" : "bg-black/20"
-            }`}
-            title="Take full quality screenshot (PNG)"
-          >
-            {isCapturing ? (
-              <span className="animate-pulse">⌛</span>
-            ) : (
-              <span>📸</span>
-            )}
-          </button>
-
-          <button
-            onClick={downloadPdf}
-            disabled={isPdfGenerating || status !== "streaming"}
-            className={`flex h-6 w-6 items-center justify-center rounded-md border border-border-subtle transition-colors hover:bg-white/5 disabled:opacity-50 ${
-              isPdfGenerating ? "bg-primary/20" : "bg-black/20"
-            }`}
-            title="Download as PDF"
-          >
-            {isPdfGenerating ? (
-              <span className="animate-pulse">⌛</span>
-            ) : (
-              <span>📄</span>
-            )}
-          </button>
-        </div>
-        <span>{sizeDisplay}</span>
-
-      </div>
+      <ScreencastStatusBar
+        tabId={tabId}
+        status={status}
+        localFps={localFps}
+        setLocalFps={setLocalFps}
+        fpsDisplay={fpsDisplay}
+        sizeDisplay={sizeDisplay}
+      />
     </div>
   );
 }
